A bit of ldt.sh cleanup
I did quite a bit of work on this, then reverted it all because it
was getting too complicated.  I don't want ldt.sh to get complicated.
I would like the same ldt.sh to be able to build for Linux and
Windows, though, so I'll have to hack it some more to get that
working.