Added a map file for testing, renamed entities
Entities were too general-purpose.  I need to stop generalizing and
make things more specific, so I renamed them "actor" and removed the
is_animal and is_alive elements.  All actors will be moving things
which are alive.

I need to get map loading and saving working - preferably today.
The first release of this game will be tomorrow, so I _need_ to get
something worthy of showing to people done, even if it's only
slightly technically interesting.  Loading and saving might work
for that.

Also, to keep myself organized, I made a TODO file with all the
things I know I need to do in it.  First is saving, because that's
important.