Updated ldtwin.sh
I really need to merge the two versions of ldt.sh into a single
script.  At this point, the only difference is a few variables; I
could have them one script that takes a flag to change the values
of those variables.  I tried to do that earlier, but for some
reason I gave up on those efforts.